Redesign notifier API
For the long period notifier was used just for reporting on exceptions.
Our API was presented mainly by the signature of the NotifyAsync method.
It was designed to accept Exception, IHttpContext and Severity. This is
an unusual combination of parameters as, typically, loggers accept log
level (severity in our case), exception and error message with some kind
of the formatter or objects to fill the placeholders. As we didn't
support error messages our variant was Okay. But now we support error
messages. Moreover, we support variadic parameter for message template
properties (similar to string.Format method from the standard library)
and that parameter must be the last one. Problem is with IHttpContext
parameter. It doesn't fit here anymore. And the only one way to solve it
gracefully is to redesign the API. One option is to split the phases of
building notice and sending it. In that case, we get access to the
notice object that should allow us to attach the HTTP context to it as
well as support for any other logic for the notice that might be needed
in the future without breaking the API again.